separate timer configs from motor_hardware.c
